requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.53.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
python3-openid==3.2.0
replicate==0.6.1
requests==2.28.2
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.2
yarl==1.8.2
